+++ docs/components/Header.tsx
@@ -7,9 +7,7 @@ import Link from 'next/link';
 import { useCallback } from 'react';
 import { useMediaQuery } from '../lib/media';
 
-/* SEARCH COMING SOON */
-/* import { SearchField } from './primitives/SearchField'; */
-
+import { SearchField } from './primitives/SearchField';
 import { Highlight } from './primitives/Highlight';
 import { Wrapper } from './primitives/Wrapper';
 import { Hamburger } from './icons/Hamburger';
@@ -19,9 +17,8 @@ import { DarkModeBtn } from './DarkModeBtn';
 import { Keystone } from './icons/Keystone';
 import { MobileMenu } from './MobileMenu';
 import { GitHub } from './icons/GitHub';
-
-/* SEARCH COMING SOON */
-/* import { Search } from './icons/Search'; */
+// TODO: Add in search for mobile via this button
+// import { Search } from './icons/Search';
 
 type HeaderContextType = { mobileNavIsOpen: boolean };
 const HeaderContext = createContext<HeaderContextType>({ mobileNavIsOpen: false });
@@ -33,8 +30,9 @@ function Logo() {
   return (
     <div
       css={mq({
-        marginRight: [0, null, null, null, '2rem'],
+        marginRight: [0, null, null, null, '1rem'],
         marginTop: '0.1rem',
+        whiteSpace: 'nowrap',
       })}
     >
       <Link href="/" passHref>
@@ -94,7 +92,13 @@ function LinkItem({ children, href }: { children: ReactNode; href: string }) {
 
   return (
     <span css={mq({ display: ['none', 'inline'], fontWeight: 600 })}>
-      <NavItem isActive={isActive} href={href} css={mq({ padding: ['0', null, '0 0.2rem'] })}>
+      <NavItem
+        isActive={isActive}
+        href={href}
+        css={{
+          padding: '0 !important',
+        }}
+      >
         {children}
       </NavItem>
     </span>
@@ -110,6 +114,44 @@ export function Header() {
 
   useEffect(() => {
     document.body.style.overflow = 'auto';
+    // search - init field
+    let searchAttempt = 0;
+    // @ts-ignore
+    document.getElementById('search-field').disabled = true;
+    const loadSearch = (searchAttempt: number) => {
+      // @ts-ignore
+      if (window.docsearch && searchAttempt < 10) {
+        // @ts-ignore
+        document.getElementById('search-field').disabled = false;
+        // @ts-ignore
+        window.docsearch({
+          apiKey: '211e94c001e6b4c6744ae72fb252eaba',
+          indexName: 'keystonejs',
+          inputSelector: '#search-field',
+        });
+      } else if (searchAttempt >= 10) {
+        // @ts-ignore
+        document.getElementById('search-field-container').style.visibility = 'hidden';
+      } else {
+        setTimeout(() => loadSearch(searchAttempt++), 500);
+      }
+    };
+    loadSearch(searchAttempt);
+    // search - keyboard shortcut
+    let keysPressed = {};
+    document.body.addEventListener('keydown', event => {
+      // @ts-ignore
+      keysPressed[event.key] = true;
+      // @ts-ignore
+      if (keysPressed['Meta'] && event.key == 'k') {
+        event.preventDefault();
+        document.getElementById('search-field')?.focus();
+      }
+    });
+    document.body.addEventListener('keyup', event => {
+      // @ts-ignore
+      delete keysPressed[event.key];
+    });
   }, []);
 
   const handleOpen = useCallback(() => {
@@ -137,19 +179,19 @@ export function Header() {
         css={mq({
           display: 'grid',
           gridTemplateColumns: [
-            /* SEARCH COMING SOON */
-            /* 'auto max-content max-content max-content',
+            'auto max-content max-content max-content',
             'auto max-content max-content max-content max-content max-content max-content',
             'max-content auto max-content max-content max-content max-content max-content',
             'max-content auto max-content max-content max-content max-content max-content max-content',
-            '15rem auto max-content max-content max-content max-content max-content max-content', */
-            'auto max-content max-content max-content',
-            'auto max-content max-content max-content max-content max-content max-content',
-            'auto max-content max-content max-content max-content max-content',
-            'auto max-content max-content max-content max-content max-content max-content',
-            'auto max-content max-content max-content max-content max-content max-content',
+            '15rem auto max-content max-content max-content max-content max-content max-content',
           ],
-          gap: ['var(--space-medium)', null, null, 'var(--space-large)', 'var(--space-xlarge)'],
+          gap: [
+            'var(--space-medium)',
+            'var(--space-large)',
+            'var(--space-medium)',
+            'var(--space-large)',
+            'var(--space-xlarge)',
+          ],
           justifyItems: 'start',
           alignItems: 'center',
           paddingTop: 'var(--space-xlarge)',
@@ -160,26 +202,25 @@ export function Header() {
           },
         })}
       >
-        <div>
-          <Logo />
-        </div>
+        <Logo />
 
-        {/* SEARCH COMING SOON */}
-        {/* <div
+        <div
+          id="search-field-container"
           css={mq({
             display: ['none', null, 'block'],
             width: ['100%', null, null, null, '80%'],
           })}
         >
           <SearchField />
-        </div> */}
+        </div>
 
         <LinkItem href="/why-keystone">Why Keystone</LinkItem>
         <LinkItem href="/updates">Updates</LinkItem>
         <LinkItem href="/docs">Docs</LinkItem>
 
-        {/* SEARCH COMING SOON */}
-        {/* <button
+        {/* TODO: Add in search for mobile via this button */}
+        {/*
+        <button
           css={mq({
             display: ['inline-block', 'inline-block', 'none'],
             appearance: 'none',
@@ -192,7 +233,8 @@ export function Header() {
           })}
         >
           <Search css={{ height: '1.4rem', marginTop: '0.2rem' }} />
-        </button> */}
+        </button>
+        */}
 
         <DarkModeBtn />
         <Button
@@ -226,13 +268,17 @@ export function Header() {
           <GitHub css={{ height: '1.5em' }} />
         </a>
         <HeaderContext.Provider value={{ mobileNavIsOpen }}>
-          <div ref={menuRef}>
+          <div
+            ref={menuRef}
+            css={mq({
+              display: ['inline-block', null, 'none'],
+            })}
+          >
             <button
               onClick={handleOpen}
               id="skip-link-navigation-btn"
               tabIndex={0}
               css={mq({
-                display: ['inline-block', null, 'none'],
                 appearance: 'none',
                 border: '0 none',
                 boxShadow: 'none',