+++ examples/document-field/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @keystone-next/example-document-field
 
+## 1.1.0
+
+### Minor Changes
+
+- [#6038](https://github.com/keystonejs/keystone/pull/6038) [`55d50573d`](https://github.com/keystonejs/keystone/commit/55d50573dd77c615c7f9a31c2bc649a9261ab24d) Thanks [@timleslie](https://github.com/timleslie)! - Added inline relationship example.
+
+### Patch Changes
+
+- Updated dependencies [[`38b78f2ae`](https://github.com/keystonejs/keystone/commit/38b78f2aeaf4c5d8176a1751ad8cb5a7acce2790), [`139d7a8de`](https://github.com/keystonejs/keystone/commit/139d7a8def263d40c0d1d5353d2744842d9a0951), [`279403cb0`](https://github.com/keystonejs/keystone/commit/279403cb0b4bffb946763c9a7ef71be57478eeb3), [`890e3d0a5`](https://github.com/keystonejs/keystone/commit/890e3d0a500ecc30cc88946ba53438812b11b2a4), [`253df44c2`](https://github.com/keystonejs/keystone/commit/253df44c2f8d6535a6425b2593eaed5380433d57), [`253df44c2`](https://github.com/keystonejs/keystone/commit/253df44c2f8d6535a6425b2593eaed5380433d57), [`f482db633`](https://github.com/keystonejs/keystone/commit/f482db6332e54a1d5cd469e2805b99b544208e83), [`093f9791b`](https://github.com/keystonejs/keystone/commit/093f9791bc37357b9700a9c49a7ae1102757bca5), [`890e3d0a5`](https://github.com/keystonejs/keystone/commit/890e3d0a500ecc30cc88946ba53438812b11b2a4), [`c536b478f`](https://github.com/keystonejs/keystone/commit/c536b478fc89f2d933cddf8533e7d88030540a63)]:
+  - @keystone-next/fields@12.0.0
+  - @keystone-next/keystone@22.0.0
+  - @keystone-next/document-renderer@4.0.0
+  - @keystone-next/fields-document@7.0.2
+
 ## 1.0.2
 
 ### Patch Changes